renamed module for consistency
